Listener integrated with Nethereum.Web3

+++ BlockInfo.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 
 namespace WebJobs.Extensions.Web3.BlockTrigger
 {
     public class BlockInfo
     {
-        public int Height { get; set; }
+        public BigInteger Height { get; set; }
         public string Hash { get; set; }
     }
 }
